+++ cookbooks/aws-parallelcluster-slurm/attributes/versions.rb
@@ -1,10 +1,10 @@
 # Slurm
-default['cluster']['slurm']['version'] = '23-11-4-1'
+default['cluster']['slurm']['version'] = '23-11-7-1'
 default['cluster']['slurm']['commit'] = ''
 default['cluster']['slurm']['branch'] = ''
-default['cluster']['slurm']['sha256'] = '832bc076d8ac9fd44cf2bf539c5046d7aa3ec20fb6c699693552dce3e5ad5588'
+default['cluster']['slurm']['sha256'] = 'b25127efd69a47c14bd65dfa3bff2687b5350c5290eafb601f923faebe6fd238'
 default['cluster']['slurm']['base_url'] = "https://github.com/SchedMD/slurm/archive"
 # Munge
-default['cluster']['munge']['munge_version'] = '0.5.15'
-default['cluster']['munge']['sha256'] = '51b2c81d1a7ec2ab5d486fa51b50c7e79eb1810ca6687b6ed65f3601abc55614'
+default['cluster']['munge']['munge_version'] = '0.5.16'
+default['cluster']['munge']['sha256'] = 'fa27205d6d29ce015b0d967df8f3421067d7058878e75d0d5ec3d91f4d32bb57'
 default['cluster']['munge']['base_url'] = "https://github.com/dun/munge/archive"