+++ cookbooks/third-party/selinux-6.1.12/test/cookbooks/selinux_test/recipes/enforcing.rb
@@ -1,6 +1,0 @@
-if platform_family?('debian')
-  # Ubuntu breaks kitchen SSH connections by default so need to load a module or two first
-  include_recipe 'selinux::permissive'
-else
-  include_recipe 'selinux::enforcing'
-end